+++ src/Console/Application.php
@@ -25,8 +25,10 @@ use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Contracts\Cache\ItemInterface;
+use Symfony\Contracts\HttpClient\Exception\ExceptionInterface;
 
 use function Castor\log;
+use function Castor\request;
 use function Castor\run;
 
 /** @internal */
@@ -161,19 +163,16 @@ class Application extends SymfonyApplication
     {
         $latestVersion = GlobalHelper::getCache()->get('latest-version', function (ItemInterface $item): array {
             $item->expiresAfter(3600 * 60 * 24);
-            $opts = [
-                'http' => [
-                    'method' => 'GET',
-                    'header' => 'User-Agent: castor',
-                    'timeout' => 5,
-                ],
-            ];
 
-            $context = stream_context_create($opts);
+            $response = request('GET', 'https://api.github.com/repos/jolicode/castor/releases/latest', [
+                'timeout' => 1,
+            ]);
 
-            $content = @file_get_contents('https://api.github.com/repos/jolicode/castor/releases/latest', false, $context);
-
-            return json_decode($content ?: '', true) ?? [];
+            try {
+                return $response->toArray();
+            } catch (ExceptionInterface) {
+                return [];
+            }
         });
 
         if (!$latestVersion) {