+++ examples/run.php
@@ -5,10 +5,10 @@ namespace run;
 use Castor\Attribute\AsTask;
 use Symfony\Component\Console\Helper\ProcessHelper;
 
+use function Castor\app;
 use function Castor\capture;
-use function Castor\get_application;
-use function Castor\get_exit_code;
-use function Castor\get_output;
+use function Castor\exit_code;
+use function Castor\output;
 use function Castor\run;
 
 #[AsTask(description: 'Run a sub-process and display information about it')]
@@ -46,16 +46,16 @@ function whoami(): void
 #[AsTask(description: 'Run a sub-process and return its exit code, with get_exit_code() function')]
 function testFile(): int
 {
-    return get_exit_code('test -f unknown-file');
+    return exit_code('test -f unknown-file');
 }
 
 #[AsTask(description: 'Run a sub-process and display information about it, with ProcessHelper')]
 function with_process_helper(): void
 {
-    if (!get_output()->isVeryVerbose()) {
-        get_output()->writeln('Re-run with -vv, -vvv to see the output of the process.');
+    if (!output()->isVeryVerbose()) {
+        output()->writeln('Re-run with -vv, -vvv to see the output of the process.');
     }
     /** @var ProcessHelper */
-    $helper = get_application()->getHelperSet()->get('process');
-    $helper->run(get_output(), ['ls', '-alh']);
+    $helper = app()->getHelperSet()->get('process');
+    $helper->run(output(), ['ls', '-alh']);
 }